--- a/setup.py
+++ b/setup.py
@@ -336,19 +336,17 @@ ext_modules += [
     )
 ]
 
-# sandbox
-# ext_modules += [
-#     Extension(
-#         name="mrsimulator.sandbox",
-#         sources=[*source, "src/c_lib/sandbox/sandbox" + ext],
-#         include_dirs=include_dirs,
-#         language="c",
-#         libraries=libraries,
-#         library_dirs=library_dirs,
-#         extra_compile_args=extra_compile_args,
-#         extra_link_args=extra_link_args,
-#     )
-# ]
+# c-lib
+ext_modules += [
+    Extension(
+        name="mrsimulator.clib",
+        sources=["src/c_lib/lib/histogram.c", "src/c_lib/clib/clib" + ext],
+        include_dirs=include_dirs,
+        language="c",
+        extra_compile_args=extra_compile_args,
+        extra_link_args=extra_link_args,
+    )
+]
 
 if USE_CYTHON:
     ext_modules = cythonize(ext_modules, language_level=3, gdb_debug=False)