--- a/setup.py
+++ b/setup.py
@@ -304,6 +304,7 @@ source = [
     "src/c_lib/lib/frequency_averaging.c",
     "src/c_lib/lib/schemes.c",
     "src/c_lib/lib/simulation.c",
+    "src/c_lib/lib/vm_linalg.c",
 ]
 
 ext = ".pyx" if USE_CYTHON else ".c"