--- a/src/c_lib/base/base_model.pxd
+++ b/src/c_lib/base/base_model.pxd
@@ -176,7 +176,7 @@ cdef extern from "simulation.h":
 
 
 cdef extern from "vm_linalg.h":
-    void cubic_roots(
+    void haeberlen_components(
         int n,
         double *expr_base_p,
         double *expr_base_q,