--- a/src/c_lib/base/base_model.pyx
+++ b/src/c_lib/base/base_model.pyx
@@ -609,19 +609,20 @@ def calculate_transition_connect_weight(
 @cython.profile(False)
 @cython.boundscheck(False)
 @cython.wraparound(False)
-def cubic_roots(
+def get_Haeberlen_components(
         ndarray[double, ndim=2] expr_base_p,
         ndarray[double, ndim=2] expr_base_q,
         double zeta,
         double eta,
         double rho
     ):
+    """Return random extended czjzek tensors in Haeberlen convention""""
     cdef int n = expr_base_p.shape[1]
     cdef ndarray[double, ndim=2] param = np.empty((n, 2), dtype=float)
-    clib.cubic_roots(
+    clib.haeberlen_components(
         n, &expr_base_p[0, 0], &expr_base_q[0, 0], zeta, eta, rho, &param[0, 0]
     )
-    return param
+    return param[:, 0], param[:, 1]
 
 # @cython.profile(False)
 # @cython.boundscheck(False)