--- a/src/c_lib/base/base_model.pyx
+++ b/src/c_lib/base/base_model.pyx
@@ -614,15 +614,15 @@ def get_Haeberlen_components(
         ndarray[double, ndim=2] expr_base_q,
         double zeta,
         double eta,
-        double rho
-    ):
-    """Return random extended czjzek tensors in Haeberlen convention""""
+        double rho):
+    """Return random extended czjzek tensors in Haeberlen convention"""
     cdef int n = expr_base_p.shape[1]
     cdef ndarray[double, ndim=2] param = np.empty((n, 2), dtype=float)
-    clib.haeberlen_components(
+    clib.vm_haeberlen_components(
         n, &expr_base_p[0, 0], &expr_base_q[0, 0], zeta, eta, rho, &param[0, 0]
     )
     return param[:, 0], param[:, 1]
+
 
 # @cython.profile(False)
 # @cython.boundscheck(False)