--- a/src/c_lib/include/config.h
+++ b/src/c_lib/include/config.h
@@ -26,9 +26,10 @@ typedef float complex64[2];
 #define CONST_iPI CONST_2PI *I
 #define TOL 1.0e-6
 
-#define modd(x, y) ((x) - (int)((x) / (y)) * (y))  // fold x within range y
-#define lerp(w, v1, v2) ((1.0 - (w)) * (v1) + (w) * (v2))
-#define sign(x) (int)(((x) > 0) - ((x) < 0))  // return sign of x
+// modd = modulus function
+#define modd(x, y) ((x) - (int)((x) / (y)) * (y))          // fold x within range y
+#define lerp(w, v1, v2) ((1.0 - (w)) * (v1) + (w) * (v2))  // linear interpolation
+#define sign(x) (int)(((x) > 0) - ((x) < 0))               // return sign of x
 
 // Compiler version check
 #if __STDC_VERSION__ >= 199901L