--- a/src/c_lib/include/vm_linalg.h
+++ b/src/c_lib/include/vm_linalg.h
@@ -1,0 +1,10 @@
+
+#include "config.h"
+
+#ifndef __v_linalg__
+#define __v_linalg__
+
+void cubic_roots(int n, double *expr_base_p, double *expr_base_q, double zeta,
+                 double eta, double rho, double *param);
+
+#endif /* __vmath__ */