--- a/src/c_lib/include/vm_linalg.h
+++ b/src/c_lib/include/vm_linalg.h
@@ -4,7 +4,7 @@
 #ifndef __v_linalg__
 #define __v_linalg__
 
-void cubic_roots(int n, double *expr_base_p, double *expr_base_q, double zeta,
-                 double eta, double rho, double *param);
+void haeberlen_components(int n, double *expr_base_p, double *expr_base_q, double zeta,
+                          double eta, double rho, double *param);
 
 #endif /* __vmath__ */