--- a/src/c_lib/lib/histogram.c
+++ b/src/c_lib/lib/histogram.c
@@ -9,6 +9,7 @@
 
 #include "histogram.h"
 
+#include <stdbool.h>
 #include <stdio.h>
 
 void histogram1d_c(int x_count, const double x_min, const double x_max,
@@ -62,34 +63,57 @@ void histogram2d_c(int x_count, const double x_min, const double x_max, int y_co
   }
 }
 
-// void histogram2d_linterp(int x_count, const double hist_x_min, const double
-// hist_x_max,
-//                  int y_count, const double hist_y_min, const double hist_y_max,
-//                  double *restrict hist, int sample_count,
-//                  const double *restrict sample_x, const double *restrict sample_y,
-//                  const double *restrict weights) {
-//   int i, ix, iy, hist_index;
-//   double temp_sample_x, temp_sample_y, normx, normy;
-//   double interval_x, interval_y;
+void histogram2d_interp_c(int x_count, const double x_min, const double x_max,
+                          int y_count, const double y_min, const double y_max,
+                          double *restrict hist, int sample_count,
+                          const double *restrict sample_x, const int stride_x,
+                          const double *restrict sample_y, const int stride_y,
+                          const double *restrict weights, const int stride_w) {
+  int ix, iy, p, counter = sample_count;
+  bool left;
+  double temp_sample_x, temp_sample_y, normx, normy;
+  double *x_pt = sample_x, *y_pt = sample_y, *w_pt = weights;
+  double delta_x, delta_y, x_min_adj, y_min_adj;
 
-//   interval_x = (hist_x_max - hist_x_min) / (x_count - 1);
-//   interval_y = (hist_y_max - hist_y_min) / (y_count - 1);
+  normx = (double)(x_count) / (x_max - x_min);
+  normy = (double)(y_count) / (y_max - y_min);
 
-//   normx = 1.0 / interval_x;
-//   normy = 1.0 / interval_y;
+  x_min_adj = x_min;  // - 1.0 / normx;
+  y_min_adj = y_min;  // - 1.0 / normy;
 
-//   for (i = 0; i <= sample_count - 1; i++) {
-//     temp_sample_x = sample_x[i];
-//     temp_sample_y = sample_y[i];
+  while (counter-- > 0) {
+    temp_sample_x = *x_pt;
+    if (temp_sample_x >= x_min_adj && temp_sample_x < x_max) {
+      temp_sample_y = *y_pt;
+      if (temp_sample_y >= y_min_adj && temp_sample_y < y_max) {
+        ix = (int)((temp_sample_x - x_min) * normx);
+        iy = (int)((temp_sample_y - y_min) * normy);
+        p = iy + y_count * ix;
 
-//     if (temp_sample_x >= hist_x_min &&
-//         temp_sample_x < hist_x_max &&
-//         temp_sample_y >= hist_y_min &&
-//         temp_sample_y < hist_y_max) {
-//       ix = (int)((temp_sample_x - hist_x_min) * normx);
-//       iy = (int)((temp_sample_y - hist_y_min) * normy);
-//       hist_index = iy + y_count * ix;
-//       hist[hist_index] += weights[i];
-//     }
-//   }
-// }
+        delta_y = temp_sample_y - iy / normy;
+        delta_x = temp_sample_x - ix / normx;
+
+        left = (delta_y) < 0;
+        if (left) {
+          if (iy != 0) hist[p - 1] -= *w_pt * delta_y;
+          hist[p] += *w_pt * (1.0 + delta_y);
+        } else {
+          if (iy != y_count) hist[p + 1] += *w_pt * delta_y;
+          hist[p] += *w_pt * (1.0 - delta_y);
+        }
+
+        left = (delta_x) < 0;
+        if (left) {
+          if (iy != 0) hist[p - y_count] -= *w_pt * delta_x;
+          hist[p] += *w_pt * (1.0 + delta_x);
+        } else {
+          if (iy != y_count) hist[p + y_count] += *w_pt * delta_x;
+          hist[p] += *w_pt * (1.0 - delta_x);
+        }
+      }
+    }
+    x_pt += stride_x;
+    y_pt += stride_y;
+    w_pt += stride_w;
+  }
+}