- Optimize extended czjzek distribution sampling `0.4323 s -> 0.0427 s`
- Optimize 1D histogram `0.08032 s -> 0.002789 s`
- Optimize 2D histogram
<img width="508" alt="Screenshot 2024-04-04 at 8 26 11 PM" src="https://github.com/deepanshs/mrsimulator/assets/21365911/af345f94-5061-49bc-82bf-fa744fefbe62">